コメントを実装 整数リテラルの読み込みに使う関数をstoiからstoulにした

+++ hbs_value__read_integer.cpp
@@ -12,7 +12,7 @@ read_decimal_literal(const mkf::Node&  base)
 
   base.collect_characters(s);
 
-  return std::stoi(s,nullptr,10);
+  return static_cast<int>(std::stoul(s,nullptr,10));
 }
